+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @0xsodium/auth
 
+## 1.14.0
+
+### Minor Changes
+
+- Support web iframe app transports
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/abi@1.14.0
+  - @0xsodium/config@1.14.0
+  - @0xsodium/network@1.14.0
+  - @0xsodium/utils@1.14.0
+  - @0xsodium/wallet@1.14.0
+
 ## 1.13.0
 
 ### Minor Changes